+++ eslint-configs/eslint-config-seekingalpha-typescript/CHANGELOG.md
@@ -1,4 +1,18 @@
 # Change Log
 
+## 1.0.1 - 2022-11-09
+ - [breaking] update `@typescript-eslint/space-before-function-paren` rule preferences
+ - [breaking] update `@typescript-eslint/brace-style` rule preferences
+ - [breaking] update `@typescript-eslint/comma-spacing` rule preferences
+ - [breaking] update `@typescript-eslint/dot-notation` rule preferences
+ - [breaking] update `@typescript-eslint/func-call-spacing` rule preferences
+ - [breaking] update `@typescript-eslint/keyword-spacing` rule preferences
+ - [breaking] update `@typescript-eslint/lines-between-class-members` rule preferences
+ - [breaking] update `@typescript-eslint/no-shadow` rule preferences
+ - [breaking] update `@typescript-eslint/no-use-before-define` rule preferences
+ - [breaking] update `@typescript-eslint/padding-line-between-statements` rule preferences
+ - [breaking] update `@typescript-eslint/semi` rule preferences
+ - [breaking] update `@typescript-eslint/no-extra-parens` rule preferences
+
 ## 1.0.0 - 2022-11-08
  - Initial commit